+++ mokakbob-core/src/main/java/com/mokakbob/common/exception/DomainErrorCode.java
@@ -1,4 +1,4 @@
-package com.mokakbob.common.exception.exceptions;
+package com.mokakbob.common.exception;
 
 public interface DomainErrorCode {
