+++ mokakbob-core/src/main/java/com/mokakbob/common/exception/DomainException.java
@@ -1,4 +1,4 @@
-package com.mokakbob.common.exception.exceptions;
+package com.mokakbob.common.exception;
 
 public class DomainException extends RuntimeException {
